* commit '7fc9d908cf84e562da459e7fa1beaa2892c23526': (74 commits)
  NTR - Fix Phpstan errors
  SW-23856 - Open shipping cost calculation in basket on reload
  SW-23809 - Add checks for valid ids in Listing and Blog controllers
  SW-23771 - Modify help text
  NTR - Fix Request type in Criteria objects
  SW-23880 - Standardize on one JS CDN, add resource integrity
  SW-23892 - Removes obsolete condition for currency factor and adds test with currency factors for sBasket::sAddVoucher() fixes #2075
  SW-23882 - SW-23873 Fix manufacturer sorting fixes #2084
  SW-23813 - Fix snippets
  NTR - Update PHPStan and increase level to 5
  NTR - Update Symfony to 3.4.24
  SW-23778 - Fix media manager blog album selection
  SW-23768 - Fix race condition on confirm page
  SW-23820 - hide non-sense informations in configCache for Redis
  NTR - Change SnippetManager to consider plugin directories and theme directories in readFromIni mode
  NTR - Change SnippetManager to consider plugin directories and theme directories in readFromIni mode
  SW-23858 - Changed LIKE search for ordernumber to use the index on the table. This dramatically decreases the search time and also corrects the placeholder search of ordernumbers. fixes #2080
  SW-23819 - Always activate plugin although it might be installed already fixes #2069
  SW-23813 - Add pageSize dropdown to variants list fixes #2070
  Add multi delete button to toolbar
  ...

+++ engine/Shopware/Components/Test/CheckoutTest.php
@@ -24,14 +24,43 @@
 
 namespace Shopware\Components\Test;
 
+use Enlight_Controller_Request_RequestHttp;
 use Shopware\Components\Random;
+use Shopware\Models\Shop\Shop;
+use Shopware\Tests\Functional\Bundle\StoreFrontBundle\Helper;
 
 abstract class CheckoutTest extends \Enlight_Components_Test_Controller_TestCase
 {
+    const USER_AGENT = 'Mozilla/5.0 (Android; Tablet; rv:14.0) Gecko/14.0 Firefox/14.0';
+
+    public $clearBasketOnReset = true;
+
+    /**
+     * @var Helper
+     */
+    protected $apiHelper;
+
+    public function setUp()
+    {
+        parent::setUp();
+        $this->apiHelper = new Helper();
+    }
+
+    protected function tearDown()
+    {
+        parent::tearDown();
+        $this->apiHelper->cleanUp();
+    }
+
     public function reset()
     {
         parent::reset();
-        Shopware()->Container()->get('dbal_connection')->executeQuery('DELETE FROM s_order_basket');
+
+        if ($this->clearBasketOnReset) {
+            Shopware()->Container()->get('dbal_connection')->executeQuery('DELETE FROM s_order_basket');
+        }
+
+        $this->Request()->setHeader('User-Agent', self::USER_AGENT);
     }
 
     /**
@@ -42,14 +71,14 @@ abstract class CheckoutTest extends \Enlight_Components_Test_Controller_TestCase
      */
     protected function createArticle($price = 10, $taxRate = 19.0)
     {
-        $resourceHelper = new \Shopware\Tests\Functional\Bundle\StoreFrontBundle\Helper();
-        $orderNumber = 'swTEST' . uniqid(rand());
+        $orderNumber = 'swTEST' . uniqid((string) rand());
 
-        $resourceHelper->createArticle([
+        $this->apiHelper->createArticle([
             'name' => 'Testartikel',
             'description' => 'Test description',
             'active' => true,
             'mainDetail' => [
+                'active' => true,
                 'number' => $orderNumber,
                 'inStock' => 15,
                 'lastStock' => true,
@@ -73,6 +102,23 @@ abstract class CheckoutTest extends \Enlight_Components_Test_Controller_TestCase
         ]);
 
         return $orderNumber;
+    }
+
+    protected function updateProductPrice($orderNumber, $price, $taxRate)
+    {
+        $this->apiHelper->updateArticle($orderNumber, [
+            'mainDetail' => [
+                'prices' => [
+                    [
+                        'customerGroupKey' => 'EK',
+                        'from' => 1,
+                        'to' => '-',
+                        'price' => $price,
+                    ],
+                ],
+            ],
+            'tax' => $taxRate,
+        ]);
     }
 
     /**
@@ -189,5 +235,60 @@ abstract class CheckoutTest extends \Enlight_Components_Test_Controller_TestCase
         ], [
             'ordercode' => $orderCode,
         ]);
+    }
+
+    /**
+     * Login as a frontend user
+     */
+    protected function loginFrontendUser()
+    {
+        Shopware()->Front()->setRequest(new Enlight_Controller_Request_RequestHttp());
+        $user = Shopware()->Db()->fetchRow(
+            'SELECT id, email, password, subshopID, language FROM s_user WHERE id = 1'
+        );
+
+        /** @var \Shopware\Models\Shop\Repository $repository */
+        $repository = Shopware()->Models()->getRepository(Shop::class);
+        $shop = $repository->getActiveById($user['language']);
+
+        $shop->registerResources();
+
+        Shopware()->Session()->Admin = true;
+        Shopware()->System()->_POST = [
+            'email' => $user['email'],
+            'passwordMD5' => $user['password'],
+        ];
+        Shopware()->Modules()->Admin()->sLogin(true);
+    }
+
+    protected function addProduct($productNumber, $quantity = 1)
+    {
+        $this->reset();
+        $this->Request()->setMethod('POST');
+        $this->Request()->setHeader('User-Agent', self::USER_AGENT);
+        $this->Request()->setParam('sQuantity', $quantity);
+        $this->Request()->setParam('sAdd', $productNumber);
+        $this->dispatch('/checkout/addArticle');
+    }
+
+    protected function visitCart()
+    {
+        $this->reset();
+        $this->dispatch('/checkout/cart');
+    }
+
+    protected function visitConfirm()
+    {
+        $this->reset();
+        $this->Request()->setMethod('POST');
+        $this->dispatch('/checkout/confirm');
+    }
+
+    protected function visitFinish()
+    {
+        $this->reset();
+        $this->Request()->setMethod('POST');
+        $this->Request()->setParam('sAGB', 'on');
+        $this->dispatch('/checkout/finish');
     }
 }
